Updating limb darkening API for better portability use (i.e. CITISENS). Also, found a duplication in code that was removed and added in wheel library into requirements.txt.
